// src/app/pp-core/local-storage/local-storage-items.enum.ts
export enum LocalStorageItems {
    promotionFilters = "promotionFilters",
    promotionFiltersAuthor = "promotionFiltersAuthor",
    configSettings = "configSettings",
    user = "user",
    promotionTypes = "promotionTypes",
    promotionStatus = "promotionStatus",
    availableUnitsOfMeasurement = "availableUnitsOfMeasurement",
    selectedUnitOfMeasurement = "selectedUnitOfMeasurement",
    campaigns = "campaigns",
    campaignFilters = "camapignFilters"
}